refactor(slice): extract transition orchestrator out of cli adapter
cli command is now a thin adapter: parse flags, open stores, call
transitionSliceOrchestrator, serialize response. all orchestration
(resolve slice, precondition check, pre-stage state and checkpoint,
withTransaction plus TOCTOU re-check, response shaping) lives in the
new src/application/slice/transition-slice.ts module.

no behavior change: integration specs remain unchanged and pass.
