+++ .tff-cc/STATE.md
@@ -1,0 +1,10 @@
+# State — M
+
+## Progress
+- Slices: 0/1 completed
+- Tasks: 0/0 completed
+
+## Slices
+| Slice | Status | Tasks | Progress |
+|---|---|---|---|
+| Test Slice | researching | 0/0 | 0% |
