+++ .tff-cc/milestones/M01/slices/M01-S01/CHECKPOINT.md
@@ -1,0 +1,8 @@
+# Checkpoint — M01-S01
+- Base commit: 
+- Current wave: 0
+- Completed waves: []
+- Completed tasks: []
+- Executor log: 
+
+<!-- checkpoint-json: {"sliceId":"M01-S01","baseCommit":"","currentWave":0,"completedWaves":[],"completedTasks":[],"executorLog":[]} -->
